requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorthief==0.2.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et_xmlfile==2.0.0
Flask==3.0.3
flask-cors==5.0.1
fonttools==4.56.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.29.0
nest-asyncio==1.6.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
plotly==6.0.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyogrio==0.10.0
pyparsing==3.2.1
pypdfium2==4.30.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
shapely==2.0.7
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.43.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.0.6
zipp==3.21.0
